:art: #1497 微信公众号群发图文消息接口参数优化，保持跟普通发送接口一致，减少重复代码
