+++ weixin-java-mp/src/main/java/me/chanjar/weixin/mp/util/json/WxMpMassNewsGsonAdapter.java
@@ -2,6 +2,7 @@ package me.chanjar.weixin.mp.util.json;
 
 import com.google.gson.*;
 import me.chanjar.weixin.mp.bean.WxMpMassNews;
+import me.chanjar.weixin.mp.bean.material.WxMpNewsArticle;
 
 import java.lang.reflect.Type;
 
@@ -12,8 +13,8 @@ public class WxMpMassNewsGsonAdapter implements JsonSerializer<WxMpMassNews>, Js
     JsonObject newsJson = new JsonObject();
 
     JsonArray articleJsonArray = new JsonArray();
-    for (WxMpMassNews.WxMpMassNewsArticle article : message.getArticles()) {
-      JsonObject articleJson = WxMpGsonBuilder.create().toJsonTree(article, WxMpMassNews.WxMpMassNewsArticle.class).getAsJsonObject();
+    for (WxMpNewsArticle article : message.getArticles()) {
+      JsonObject articleJson = WxMpGsonBuilder.create().toJsonTree(article, WxMpNewsArticle.class).getAsJsonObject();
       articleJsonArray.add(articleJson);
     }
     newsJson.add("articles", articleJsonArray);
@@ -29,7 +30,7 @@ public class WxMpMassNewsGsonAdapter implements JsonSerializer<WxMpMassNews>, Js
       JsonArray articles = json.getAsJsonArray("articles");
       for (JsonElement article1 : articles) {
         JsonObject articleInfo = article1.getAsJsonObject();
-        WxMpMassNews.WxMpMassNewsArticle article = WxMpGsonBuilder.create().fromJson(articleInfo, WxMpMassNews.WxMpMassNewsArticle.class);
+        WxMpNewsArticle article = WxMpGsonBuilder.create().fromJson(articleInfo, WxMpNewsArticle.class);
         wxMpMassNews.addArticle(article);
       }
     }
