+++ weixin-java-mp/src/main/java/me/chanjar/weixin/mp/util/json/WxMpNewsArticleGsonAdapter.java
@@ -2,7 +2,7 @@ package me.chanjar.weixin.mp.util.json;
 
 import com.google.gson.*;
 import me.chanjar.weixin.common.util.json.GsonHelper;
-import me.chanjar.weixin.mp.bean.WxMpMassNews;
+import me.chanjar.weixin.mp.bean.material.WxMpNewsArticle;
 import org.apache.commons.lang3.BooleanUtils;
 
 import java.lang.reflect.Type;
@@ -10,13 +10,13 @@ import java.lang.reflect.Type;
 /**
  * @author codepiano
  */
-public class WxMpMassNewsArticleGsonAdapter implements JsonSerializer<WxMpMassNews.WxMpMassNewsArticle>, JsonDeserializer<WxMpMassNews.WxMpMassNewsArticle> {
-
+public class WxMpNewsArticleGsonAdapter implements JsonSerializer<WxMpNewsArticle>, JsonDeserializer<WxMpNewsArticle> {
   @Override
-  public JsonElement serialize(WxMpMassNews.WxMpMassNewsArticle article, Type typeOfSrc, JsonSerializationContext context) {
+  public JsonElement serialize(WxMpNewsArticle article, Type typeOfSrc, JsonSerializationContext context) {
     JsonObject articleJson = new JsonObject();
 
     articleJson.addProperty("thumb_media_id", article.getThumbMediaId());
+    articleJson.addProperty("thumb_url", article.getThumbUrl());
     articleJson.addProperty("title", article.getTitle());
     articleJson.addProperty("content", article.getContent());
     if (null != article.getAuthor()) {
@@ -29,13 +29,26 @@ public class WxMpMassNewsArticleGsonAdapter implements JsonSerializer<WxMpMassNe
       articleJson.addProperty("digest", article.getDigest());
     }
     articleJson.addProperty("show_cover_pic", article.isShowCoverPic() ? "1" : "0");
+    if (null != article.getUrl()) {
+      articleJson.addProperty("url", article.getUrl());
+    }
+
+    if (null != article.getNeedOpenComment()) {
+      articleJson.addProperty("need_open_comment",
+        BooleanUtils.toInteger(article.getNeedOpenComment(), 1, 0));
+    }
+
+    if (null != article.getOnlyFansCanComment()) {
+      articleJson.addProperty("only_fans_can_comment",
+        BooleanUtils.toInteger(article.getOnlyFansCanComment(), 1, 0));
+    }
     return articleJson;
   }
 
   @Override
-  public WxMpMassNews.WxMpMassNewsArticle deserialize(JsonElement jsonElement, Type type, JsonDeserializationContext jsonDeserializationContext) throws JsonParseException {
+  public WxMpNewsArticle deserialize(JsonElement jsonElement, Type type, JsonDeserializationContext jsonDeserializationContext) throws JsonParseException {
     JsonObject articleInfo = jsonElement.getAsJsonObject();
-    WxMpMassNews.WxMpMassNewsArticle article = new WxMpMassNews.WxMpMassNewsArticle();
+    WxMpNewsArticle article = new WxMpNewsArticle();
 
     JsonElement title = articleInfo.get("title");
     if (title != null && !title.isJsonNull()) {
@@ -61,9 +74,27 @@ public class WxMpMassNewsArticleGsonAdapter implements JsonSerializer<WxMpMassNe
     if (thumbMediaId != null && !thumbMediaId.isJsonNull()) {
       article.setThumbMediaId(GsonHelper.getAsString(thumbMediaId));
     }
+    JsonElement thumbUrl = articleInfo.get("thumb_url");
+    if (thumbUrl != null && !thumbUrl.isJsonNull()) {
+      article.setThumbUrl(GsonHelper.getAsString(thumbUrl));
+    }
     JsonElement showCoverPic = articleInfo.get("show_cover_pic");
     if (showCoverPic != null && !showCoverPic.isJsonNull()) {
       article.setShowCoverPic(BooleanUtils.toBoolean(showCoverPic.getAsInt()));
+    }
+    JsonElement url = articleInfo.get("url");
+    if (url != null && !url.isJsonNull()) {
+      article.setUrl(GsonHelper.getAsString(url));
+    }
+
+    JsonElement needOpenComment = articleInfo.get("need_open_comment");
+    if (needOpenComment != null && !needOpenComment.isJsonNull()) {
+      article.setNeedOpenComment(BooleanUtils.toBoolean(needOpenComment.getAsInt()));
+    }
+
+    JsonElement onlyFansCanComment = articleInfo.get("only_fans_can_comment");
+    if (onlyFansCanComment != null && !onlyFansCanComment.isJsonNull()) {
+      article.setOnlyFansCanComment(BooleanUtils.toBoolean(onlyFansCanComment.getAsInt()));
     }
     return article;
   }
