+++ weixin-java-mp/src/test/java/me/chanjar/weixin/mp/api/impl/WxMpMassMessageServiceImplTest.java
@@ -12,14 +12,17 @@ import me.chanjar.weixin.mp.bean.WxMpMassNews;
 import me.chanjar.weixin.mp.bean.WxMpMassOpenIdsMessage;
 import me.chanjar.weixin.mp.bean.WxMpMassTagMessage;
 import me.chanjar.weixin.mp.bean.WxMpMassVideo;
+import me.chanjar.weixin.mp.bean.material.WxMpNewsArticle;
 import me.chanjar.weixin.mp.bean.result.WxMpMassSendResult;
 import me.chanjar.weixin.mp.bean.result.WxMpMassUploadResult;
-import org.testng.annotations.*;
+import org.testng.annotations.DataProvider;
+import org.testng.annotations.Guice;
+import org.testng.annotations.Test;
 
 import java.io.IOException;
 import java.io.InputStream;
 
-import static org.testng.Assert.*;
+import static org.testng.Assert.assertNotNull;
 
 /**
  * 测试群发消息
@@ -35,15 +38,13 @@ public class WxMpMassMessageServiceImplTest {
   @Test
   public void testTextMassOpenIdsMessageSend() throws WxErrorException {
     // 发送群发消息
-    TestConfigStorage configProvider = (TestConfigStorage) this.wxService
-      .getWxMpConfigStorage();
+    TestConfigStorage configProvider = (TestConfigStorage) this.wxService      .getWxMpConfigStorage();
     WxMpMassOpenIdsMessage massMessage = new WxMpMassOpenIdsMessage();
     massMessage.setMsgType(WxConsts.MassMsgType.TEXT);
-    massMessage.setContent("测试群发消息\n欢迎欢迎，热烈欢迎\n换行测试\n超链接:<a href=\"http://www.baidu.com\">Hello World</a>");
+    massMessage.setContent("测试群发消息\n欢迎欢迎\n换行测试\n超链接:<a href=\"http://www.baidu.com\">Hello World</a>");
     massMessage.getToUsers().add(configProvider.getOpenid());
 
-    WxMpMassSendResult massResult = this.wxService.getMassMessageService()
-      .massOpenIdsMessageSend(massMessage);
+    WxMpMassSendResult massResult = this.wxService.getMassMessageService().massOpenIdsMessageSend(massMessage);
     assertNotNull(massResult);
     assertNotNull(massResult.getMsgId());
   }
@@ -51,15 +52,13 @@ public class WxMpMassMessageServiceImplTest {
   @Test(dataProvider = "massMessages")
   public void testMediaMassOpenIdsMessageSend(String massMsgType, String mediaId) throws WxErrorException {
     // 发送群发消息
-    TestConfigStorage configProvider = (TestConfigStorage) this.wxService
-      .getWxMpConfigStorage();
+    TestConfigStorage configProvider = (TestConfigStorage) this.wxService.getWxMpConfigStorage();
     WxMpMassOpenIdsMessage massMessage = new WxMpMassOpenIdsMessage();
     massMessage.setMsgType(massMsgType);
     massMessage.setMediaId(mediaId);
     massMessage.getToUsers().add(configProvider.getOpenid());
 
-    WxMpMassSendResult massResult = this.wxService.getMassMessageService()
-      .massOpenIdsMessageSend(massMessage);
+    WxMpMassSendResult massResult = this.wxService.getMassMessageService().massOpenIdsMessageSend(massMessage);
     assertNotNull(massResult);
     assertNotNull(massResult.getMsgId());
   }
@@ -68,26 +67,23 @@ public class WxMpMassMessageServiceImplTest {
   public void testTextMassGroupMessageSend() throws WxErrorException {
     WxMpMassTagMessage massMessage = new WxMpMassTagMessage();
     massMessage.setMsgType(WxConsts.MassMsgType.TEXT);
-    massMessage.setContent("测试群发消息\n欢迎欢迎，热烈欢迎\n换行测试\n超链接:<a href=\"http://www.baidu.com\">Hello World</a>");
+    massMessage.setContent("测试群发消息\n欢迎欢迎\n换行测试\n超链接:<a href=\"http://www.baidu.com\">Hello World</a>");
     massMessage
       .setTagId(this.wxService.getUserTagService().tagGet().get(0).getId());
 
-    WxMpMassSendResult massResult = this.wxService.getMassMessageService()
-      .massGroupMessageSend(massMessage);
+    WxMpMassSendResult massResult = this.wxService.getMassMessageService().massGroupMessageSend(massMessage);
     assertNotNull(massResult);
     assertNotNull(massResult.getMsgId());
   }
 
   @Test(dataProvider = "massMessages")
-  public void testMediaMassGroupMessageSend(String massMsgType, String mediaId)
-    throws WxErrorException {
+  public void testMediaMassGroupMessageSend(String massMsgType, String mediaId)    throws WxErrorException {
     WxMpMassTagMessage massMessage = new WxMpMassTagMessage();
     massMessage.setMsgType(massMsgType);
     massMessage.setMediaId(mediaId);
     massMessage.setTagId(this.wxService.getUserTagService().tagGet().get(0).getId());
 
-    WxMpMassSendResult massResult = this.wxService.getMassMessageService()
-      .massGroupMessageSend(massMessage);
+    WxMpMassSendResult massResult = this.wxService.getMassMessageService().massGroupMessageSend(massMessage);
     assertNotNull(massResult);
     assertNotNull(massResult.getMsgId());
   }
@@ -155,15 +151,15 @@ public class WxMpMassMessageServiceImplTest {
 
       // 上传图文消息
       WxMpMassNews news = new WxMpMassNews();
-      WxMpMassNews.WxMpMassNewsArticle article1 = new WxMpMassNews.WxMpMassNewsArticle();
+      WxMpNewsArticle article1 = new WxMpNewsArticle();
       article1.setTitle("标题1");
-      article1.setContent("内容1内容1内容1内容1内容1内容1内容1内容1内容1内容1内容1内容1内容1内容1内容1内容1内容1内容1内容1内容1内容1");
+      article1.setContent("内容1内容1内容1内容1内容1内容1内容1内容1内容1内容1内容1内容1内容1内容1内容1内容1内容1");
       article1.setThumbMediaId(uploadMediaRes.getMediaId());
       news.addArticle(article1);
 
-      WxMpMassNews.WxMpMassNewsArticle article2 = new WxMpMassNews.WxMpMassNewsArticle();
+      WxMpNewsArticle article2 = new WxMpNewsArticle();
       article2.setTitle("标题2");
-      article2.setContent("内容2内容2内容2内容2内容2内容2内容2内容2内容2内容2内容2内容2内容2内容2内容2内容2内容2内容2内容2内容2内容2");
+      article2.setContent("内容2内容2内容2内容2内容2内容2内容2内容2内2内容2内容2内容2内容2内容2内容2内容2内容2内容2");
       article2.setThumbMediaId(uploadMediaRes.getMediaId());
       article2.setShowCoverPic(true);
       article2.setAuthor("作者2");
@@ -183,7 +179,7 @@ public class WxMpMassMessageServiceImplTest {
 
   @Test
   public void testMassDelete() throws Exception {
-    this.wxService.getMassMessageService().delete(1L,2);
+    this.wxService.getMassMessageService().delete(1L, 2);
   }
 
 }
